WapServerConfigMock renamed appropriately
