+++ src/test/java/edu/kit/scc/dem/wapsrv/service/AbstractWapServiceTest.java
@@ -22,7 +22,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import edu.kit.scc.dem.wapsrv.app.EtagFactory;
 import edu.kit.scc.dem.wapsrv.app.EtagFactoryMock;
 import edu.kit.scc.dem.wapsrv.app.WapServerConfig;
-import edu.kit.scc.dem.wapsrv.app.WapServerConfigMock;
+import edu.kit.scc.dem.wapsrv.app.WapServerConfigSpy;
 import edu.kit.scc.dem.wapsrv.model.Container;
 import edu.kit.scc.dem.wapsrv.model.ModelFactory;
 import edu.kit.scc.dem.wapsrv.model.Page;
@@ -38,7 +38,6 @@ import edu.kit.scc.dem.wapsrv.model.validators.Validator;
 import edu.kit.scc.dem.wapsrv.model.validators.ValidatorRegistry;
 import edu.kit.scc.dem.wapsrv.repository.CollectedRepository;
 import edu.kit.scc.dem.wapsrv.repository.RepositoryMock;
-import edu.kit.scc.dem.wapsrv.repository.jena.JenaRepository;
 import edu.kit.scc.dem.wapsrv.testscommon.ModelFactoryMock;
 import org.springframework.context.annotation.ComponentScan;
 
@@ -55,7 +54,7 @@ import org.springframework.context.annotation.ComponentScan;
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = {ContainerServiceImpl.class, RepositoryMock.class, ModelFactoryMock.class,
-  EtagFactoryMock.class, JsonLdProfileRegistry.class, ValidatorRegistry.class, WapServerConfigMock.class,
+  EtagFactoryMock.class, JsonLdProfileRegistry.class, ValidatorRegistry.class, WapServerConfigSpy.class,
   JsonLdValidator.class, Validator.class})
 @ExtendWith(HoverflyExtension.class)
 @HoverflySimulate(source = @HoverflySimulate.Source(value = "w3c_simulation.json", type = HoverflySimulate.SourceType.DEFAULT_PATH))
